requirements.txt
torch>=2.2.0
transformers>=4.40.0
datasets>=2.18.0
tokenizers>=0.19.0
pyyaml>=6.0
accelerate>=0.29.0
einops>=0.7.0
numpy>=1.26.0
pytest>=8.0.0
pytest-cov>=5.0.0
